chore: global.d.ts의 props선언 각 컴포넌트로 이동
global.d.ts에 선언한 props들을 각 컴포넌트에 해당하는 props로 다시 이동시켰습니다.
